# Conflicts:
#	build.sbt
#	lock.sbt
#	src/it/scala/org/ergoplatform/it/UtxoStateNodesSyncSpec.scala
#	src/main/scala/org/ergoplatform/modifiers/history/Header.scala
#	src/main/scala/org/ergoplatform/nodeView/history/modifierprocessors/FullBlockProcessor.scala
#	src/main/scala/org/ergoplatform/nodeView/history/modifierprocessors/FullBlockPruningProcessor.scala
#	src/main/scala/org/ergoplatform/nodeView/history/modifierprocessors/ToDownloadProcessor.scala
#	src/main/scala/org/ergoplatform/nodeView/state/StateConstants.scala
#	src/main/scala/org/ergoplatform/nodeView/state/UtxoState.scala
#	src/main/scala/org/ergoplatform/settings/Constants.scala
#	src/test/scala/org/ergoplatform/nodeView/history/NonVerifyADHistorySpecification.scala
#	src/test/scala/org/ergoplatform/nodeView/state/UtxoStateSpecification.scala
#	src/test/scala/org/ergoplatform/tools/ChainGenerator.scala
#	src/test/scala/org/ergoplatform/utils/generators/ValidBlocksGenerators.scala

+++ src/test/scala/org/ergoplatform/utils/ErgoTestHelpers.scala
@@ -25,7 +25,7 @@ trait ErgoTestHelpers
 
   def await[A](f: Future[A]): A = Await.result[A](f, defaultAwaitDuration)
 
-  def updateHeight(box: ErgoBoxCandidate, creationHeight: Long): ErgoBoxCandidate =
+  def updateHeight(box: ErgoBoxCandidate, creationHeight: Int): ErgoBoxCandidate =
     new ErgoBoxCandidate(box.value, box.proposition, creationHeight, box.additionalTokens, box.additionalRegisters)
 
   def changeValue(box: ErgoBoxCandidate, delta: Long): Option[ErgoBoxCandidate] = {
